refactor: make it clean
