+++ apps/web/components/admin/project-settings.tsx
@@ -1,0 +1,143 @@
+'use client';
+
+import { FormControl, FormDescription, FormField, FormItem, FormLabel } from '@workspace/ui/components/form';
+import { Switch } from '@workspace/ui/components/switch';
+import { UseFormReturn } from 'react-hook-form';
+
+interface ProjectSettingsProps {
+  form: UseFormReturn<any>;
+}
+
+export function ProjectSettings({ form }: ProjectSettingsProps) {
+  return (
+    <div className="space-y-4">
+      <div className="grid gap-4">
+        <FormField
+          control={form.control}
+          name="isLookingForContributors"
+          render={({ field }) => (
+            <FormItem className="flex flex-row items-start space-y-0 space-x-3">
+              <FormControl>
+                <Switch checked={field.value} onCheckedChange={field.onChange} />
+              </FormControl>
+              <div className="space-y-1 leading-none">
+                <FormLabel>Looking for Contributors</FormLabel>
+                <FormDescription>
+                  Enable if the project is actively seeking contributors
+                </FormDescription>
+              </div>
+            </FormItem>
+          )}
+        />
+
+        <FormField
+          control={form.control}
+          name="isLookingForInvestors"
+          render={({ field }) => (
+            <FormItem className="flex flex-row items-start space-y-0 space-x-3">
+              <FormControl>
+                <Switch checked={field.value} onCheckedChange={field.onChange} />
+              </FormControl>
+              <div className="space-y-1 leading-none">
+                <FormLabel>Looking for Investors</FormLabel>
+                <FormDescription>
+                  Enable if the project is seeking investment
+                </FormDescription>
+              </div>
+            </FormItem>
+          )}
+        />
+
+        <FormField
+          control={form.control}
+          name="isHiring"
+          render={({ field }) => (
+            <FormItem className="flex flex-row items-start space-y-0 space-x-3">
+              <FormControl>
+                <Switch checked={field.value} onCheckedChange={field.onChange} />
+              </FormControl>
+              <div className="space-y-1 leading-none">
+                <FormLabel>Hiring</FormLabel>
+                <FormDescription>
+                  Enable if the project is actively hiring
+                </FormDescription>
+              </div>
+            </FormItem>
+          )}
+        />
+
+        <FormField
+          control={form.control}
+          name="isPublic"
+          render={({ field }) => (
+            <FormItem className="flex flex-row items-start space-y-0 space-x-3">
+              <FormControl>
+                <Switch checked={field.value} onCheckedChange={field.onChange} />
+              </FormControl>
+              <div className="space-y-1 leading-none">
+                <FormLabel>Public</FormLabel>
+                <FormDescription>
+                  Enable if the project should be publicly visible
+                </FormDescription>
+              </div>
+            </FormItem>
+          )}
+        />
+
+        <FormField
+          control={form.control}
+          name="hasBeenAcquired"
+          render={({ field }) => (
+            <FormItem className="flex flex-row items-start space-y-0 space-x-3">
+              <FormControl>
+                <Switch checked={field.value} onCheckedChange={field.onChange} />
+              </FormControl>
+              <div className="space-y-1 leading-none">
+                <FormLabel>Acquired</FormLabel>
+                <FormDescription>
+                  Enable if the project has been acquired
+                </FormDescription>
+              </div>
+            </FormItem>
+          )}
+        />
+
+        <FormField
+          control={form.control}
+          name="isPinned"
+          render={({ field }) => (
+            <FormItem className="flex flex-row items-start space-y-0 space-x-3">
+              <FormControl>
+                <Switch checked={field.value} onCheckedChange={field.onChange} />
+              </FormControl>
+              <div className="space-y-1 leading-none">
+                <FormLabel>Pinned</FormLabel>
+                <FormDescription>
+                  Enable to pin this project to the top of listings
+                </FormDescription>
+              </div>
+            </FormItem>
+          )}
+        />
+
+        <FormField
+          control={form.control}
+          name="isRepoPrivate"
+          render={({ field }) => (
+            <FormItem className="flex flex-row items-start space-y-0 space-x-3">
+              <FormControl>
+                <Switch checked={field.value} onCheckedChange={field.onChange} />
+              </FormControl>
+              <div className="space-y-1 leading-none">
+                <FormLabel>Repository Private</FormLabel>
+                <FormDescription>
+                  Enable if the repository is private
+                </FormDescription>
+              </div>
+            </FormItem>
+          )}
+        />
+      </div>
+    </div>
+  );
+}
